+++ src/main/kotlin/com/nylas/models/EmailName.kt
@@ -1,5 +1,7 @@
 package com.nylas.models
 
+import com.squareup.moshi.Json
+
 /**
  * Class representing an email address and optional name.
  */
@@ -7,9 +9,11 @@ data class EmailName(
   /**
    * Email address.
    */
+  @Json(name = "email")
   val email: String,
   /**
    * Full name.
    */
+  @Json(name = "name")
   val name: String? = null,
 )