+++ src/main/kotlin/com/nylas/models/GrantStatus.kt
@@ -1,0 +1,11 @@
+package com.nylas.models
+
+import com.squareup.moshi.Json
+
+enum class GrantStatus {
+  @Json(name = "valid")
+  VALID,
+
+  @Json(name = "invalid")
+  INVALID,
+}