+++ src/main/kotlin/com/nylas/models/MicrosoftConnectorSettings.kt
@@ -1,0 +1,14 @@
+package com.nylas.models
+
+import com.squareup.moshi.Json
+
+/**
+ * Class representing Microsoft credentials and settings.
+ */
+data class MicrosoftConnectorSettings(
+  /**
+   * The Microsoft tenant ID
+   */
+  @Json(name = "tenant")
+  val tenant: String? = null,
+)