+++ src/main/kotlin/com/nylas/models/ScheduledMessage.kt
@@ -1,5 +1,7 @@
 package com.nylas.models
 
+import com.squareup.moshi.Json
+
 /**
  * Class representing information about a scheduled message.
  */
@@ -7,13 +9,16 @@ data class ScheduledMessage(
   /**
    * The unique identifier for the scheduled message.
    */
-  val scheduleId: Int,
+  @Json(name = "schedule_id")
+  val scheduleId: String,
   /**
    * The status of the scheduled message.
    */
+  @Json(name = "status")
   val status: ScheduledMessageStatus,
   /**
    * The time the message was sent or failed to send, in epoch time.
    */
-  val closeTime: Int?,
+  @Json(name = "close_time")
+  val closeTime: Int? = null,
 )