+++ src/main/kotlin/com/nylas/models/Webhook.kt
@@ -40,7 +40,7 @@ data class Webhook(
    * The time the status field was last updated, represented as a Unix timestamp in seconds.
    */
   @Json(name = "updated_at")
-  val updatedAt: String,
+  val updatedAt: Long,
   /**
    * A human-readable description of the webhook destination.
    */
@@ -49,6 +49,6 @@ data class Webhook(
   /**
    * The email addresses that Nylas notifies when a webhook is down for a while.
    */
-  @Json(name = "notification_email_address")
-  val notificationEmailAddress: String? = null,
+  @Json(name = "notification_email_addresses")
+  val notificationEmailAddresses: List<String>? = null,
 )