+++ src/main/kotlin/com/nylas/models/WebhookWithSecret.kt
@@ -45,7 +45,7 @@ data class WebhookWithSecret(
    * The time the status field was last updated, represented as a Unix timestamp in seconds.
    */
   @Json(name = "updated_at")
-  val updatedAt: String,
+  val updatedAt: Long,
   /**
    * A human-readable description of the webhook destination.
    */
@@ -54,6 +54,6 @@ data class WebhookWithSecret(
   /**
    * The email addresses that Nylas notifies when a webhook is down for a while.
    */
-  @Json(name = "notification_email_address")
-  val notificationEmailAddress: String? = null,
+  @Json(name = "notification_email_addresses")
+  val notificationEmailAddresses: List<String>? = null,
 )