+++ src/main/kotlin/com/nylas/resources/Contacts.kt
@@ -30,7 +30,7 @@ class Contacts(client: NylasClient) : Resource<Contact>(client, Contact::class.j
   @JvmOverloads
   fun find(identifier: String, contactId: String, queryParams: FindContactQueryParams? = null): Response<Contact> {
     val path = String.format("v3/grants/%s/contacts/%s", identifier, contactId)
-    return findResource(path)
+    return findResource(path, queryParams)
   }
 
   /**
@@ -83,7 +83,7 @@ class Contacts(client: NylasClient) : Resource<Contact>(client, Contact::class.j
   @Throws(NylasApiError::class, NylasSdkTimeoutError::class)
   @JvmOverloads
   fun listGroups(identifier: String, queryParams: ListContactGroupsQueryParams? = null): ListResponse<ContactGroup> {
-    val path = String.format("v3/grants/%s/messages/contacts/groups", identifier)
+    val path = String.format("v3/grants/%s/contacts/groups", identifier)
     val responseType = Types.newParameterizedType(ListResponse::class.java, ContactGroup::class.java)
     return client.executeGet(path, responseType, queryParams)
   }