+++ src/main/kotlin/com/nylas/util/JsonHelper.kt
@@ -1,5 +1,8 @@
 package com.nylas.util
 
+import com.nylas.models.Connector.Companion.CONNECTOR_JSON_ADAPTER_FACTORY
+import com.nylas.models.CreateConnectorRequest.Companion.CREATE_CONNECTOR_JSON_ADAPTER_FACTORY
+import com.nylas.models.CreateCredentialRequest.Companion.CREATE_CREDENTIAL_JSON_ADAPTER_FACTORY
 import com.nylas.models.GetFreeBusyResponse.Companion.FREE_BUSY_JSON_FACTORY
 import com.nylas.models.When.Companion.WHEN_JSON_FACTORY
 import com.squareup.moshi.JsonAdapter
@@ -30,7 +33,7 @@ class JsonHelper {
       .add(UpdateWhenAdapter())
       .add(CreateWhenAdapter())
       .add(IMessageAdapter())
-      .add(CreateConnectorAdapter())
+      .add(UpdateConnectorAdapter())
       .add(CredentialDataAdapter())
       .add(MicrosoftAdminConsentCredentialDataAdapter())
       .add(GoogleServiceAccountCredentialDataAdapter())
@@ -38,6 +41,9 @@ class JsonHelper {
       // Polymorphic adapters
       .add(WHEN_JSON_FACTORY)
       .add(FREE_BUSY_JSON_FACTORY)
+      .add(CONNECTOR_JSON_ADAPTER_FACTORY)
+      .add(CREATE_CONNECTOR_JSON_ADAPTER_FACTORY)
+      .add(CREATE_CREDENTIAL_JSON_ADAPTER_FACTORY)
       .add(KotlinJsonAdapterFactory())
       .build()